AI/Head Pose Estimation Models/Training/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
dm-tree==0.1.8
filelock==3.13.1
flatbuffers==24.3.7
fsspec==2024.2.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.0
h5py==3.10.0
idna==3.6
Jinja2==3.1.3
keras==3.0.5
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
mtcnn==0.1.1
namex==0.0.7
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
pandas==2.2.1
pillow==10.2.0
protobuf==4.25.3
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
rich==13.7.1
scipy==1.12.0
six==1.16.0
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
torch==2.2.1+cu118
torchaudio==2.2.1+cu118
torchvision==0.17.1
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.16.0
